@@ -1,21 +1,22 @@
 {
 	"name": "char-regex",
 	"version": "1.0.2",
+	"type": "module",
 	"description": "A regex to match any full character, considering weird character ranges.",
 	"repository": "https://github.com/Richienb/char-regex.git",
 	"author": "Richie Bendall <richiebendall@gmail.com>",
 	"license": "MIT",
-	"main": "index.js",
+	"exports": "./index.js",
 	"files": [
 		"index.js",
 		"index.d.ts"
 	],
 	"engines": {
-		"node": ">=10"
+		"node": ">=12.20"
 	},
 	"scripts": {
 		"lint": "xo",
-		"test": "yarn lint && ava"
+		"test": "xo && ava && tsd"
 	},
 	"keywords": [
 		"character",
@@ -27,12 +28,10 @@
 	"dependencies": {},
 	"devDependencies": {
 		"all-chars": "^1.0.0",
-		"ava": "^3.8.2",
-		"eslint-config-richienb": "^0.4.2",
-		"xo": "^0.30.0"
-	},
-	"resolutions": {
-		"eslint": "^6.8.0"
+		"ava": "^3.15.0",
+		"eslint-config-richienb": "^0.6.2",
+		"tsd": "^0.17.0",
+		"xo": "^0.40.2"
 	},
 	"xo": {
 		"extends": "richienb"
