@@ -1,13 +1,12 @@
 /**
- * A regex to match any full character, considering weird character ranges.
- * @example
- * ```
- * const charRegex = require("char-regex");
- *
- * "❤️👊🏽".match(charRegex());
- * //=> ["❤️", "👊🏽"]
- * ```
-*/
-declare function charRegex(): RegExp
+@returns A regex to match any full character, considering weird character ranges.
 
-export = charRegex
+@example
+```
+import charRegex from "char-regex"
+
+"❤️👊🏽".match(charRegex())
+//=> ["❤️", "👊🏽"]
+```
+*/
+export default function charRegex(): RegExp
